Added a flex container on the comment card and the comment reply card. Also added an ellipsis icon on the far right of the comment card, which when opened, will show a dropdown list only for a user's own comment.
Added a javascript file in which I wrote a simple
script to handle form submission for editing replies.

A toast container will be shown - based on the response.
If successful, a success message will be shown, otherwise
an error message will be shown.

Created a simple view to handle the form submission
and updation of the comment_body.